deleting if image deletes

+++ admin/action/delete-food.php
@@ -21,12 +21,25 @@ if (isset($_GET['id']) && isset($_GET['image_name'])) {
             }else{
                 $path = "../../images/food/$image_name";
                 $remove = unlink($path);
-                $query = "DELETE FROM tbl_food WHERE id =$id";
-                $result = mysqli_query($db, $query);
+                if ($remove == true){
+                    $query = "DELETE FROM tbl_food WHERE id =$id";
+                    $result = mysqli_query($db, $query);
+    
+                    if ($result == true) {
+                        $_SESSION['deletefood'] = '<div class="alert alert-success alert-dismissible fade show p-2 w-auto d-flex h-auto align-items-center" role="alert">
+                        <strong class="mx-2">Food successfully deleted</strong>
+                        </div>';
+                        header('location: http://localhost:7882/wowfood/admin/manage-food.php');
+                    }else{
+                        $_SESSION['deletefood'] = '<div class="alert alert-success alert-dismissible fade show p-2 w-auto d-flex h-auto align-items-center" role="alert">
+                        <strong class="mx-2">Failed to remove food</strong>
+                        </div>';
+                        header('location: http://localhost:7882/wowfood/admin/manage-food.php');
+                    }
 
-                if ($result == true) {
+                }else{
                     $_SESSION['deletefood'] = '<div class="alert alert-success alert-dismissible fade show p-2 w-auto d-flex h-auto align-items-center" role="alert">
-                    <strong class="mx-2">Food successfully deleted</strong>
+                    <strong class="mx-2">Failed to remove image</strong>
                     </div>';
                     header('location: http://localhost:7882/wowfood/admin/manage-food.php');
                 }
